% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{add_var_computer}
\alias{add_var_computer}
\title{Compute the non-primary variables}
\usage{
add_var_computer(
  sim_data,
  add_vars,
  parameter_data,
  technology_variant,
  solowversion
)
}
\arguments{
\item{sim_data}{The tibble that is being filled in the simulation function.}

\item{add_vars}{Vector with the encoded variables that are secondary and need to be computed to make the simulation table complete.}

\item{parameter_data}{The output from \code{create_parameter_grid(...)}.}

\item{technology_variant}{A string indicating the exogenous (\code{technology_variant = "exo"}) or endogenous (\code{technology_variant = "endo"}) nature of technology in the respective model. A special case (\code{technology_variant = "special"}) is the technology form of endogenous technology growth (model code "ESEG").}

\item{solowversion}{The model code for the model, such as "BS" or "ESSOE". (The same variables such as WR or RR are computed differently depending on the model.)}
}
\description{
This function computes additional secondary variables that are not computed in a models respective simulation functions. The simulation functions only compute the primary variables of a model such as L, K and Y in the Basic Solow Growth Model (BS).
}
\note{
\text{See} SimulateBasicSolowModel() \text{for an example.}
}
