% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{compare_simulations}
\alias{compare_simulations}
\title{Visualise Common Variables of Different Solow Models}
\usage{
compare_simulations(simulation_list, sim_identifier_vector, vars_selection)
}
\arguments{
\item{simulation_list}{List with tibbles. The tibbles need to be the results of the simulationfunctions, e.g. \code{SimulateBasicSolowModel()}.}

\item{sim_identifier_vector}{Vector with the model codes. (The first element of \code{sim_identifier_vector} should correspond to the first element of \code{simulation_list}.)}

\item{vars_selection}{Vector with the abbreviated variable names to visualise. (This vector can of course only contain variable abbreviations that are shared by all the simulation in \code{simulation_list}.)}
}
\description{
Visualise the evolution of common variables of multiple (2 or more) Solow variants in the same graph.
}
\note{
To see which variables can be selected, that is, to see which variables are common to two models, can be seen with \code{getVariablesAvailableToBeVisualised()}.
}
\examples{
\dontrun{compare_simulations( list(SimulateBasicSolowModel(), SimulateGeneralSolowModel()), c("BS", "GS"), c("Y", "K", "L", "WR", "RR"))}
}
