% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{create_path}
\alias{create_path}
\title{Creates the a single parameter vector.}
\usage{
create_path(iv, pfc, nv, np)
}
\arguments{
\item{iv}{initial value of parameter;}

\item{pfc}{period of change
(\code{pfc = NA} is a special entry! see examples.)}

\item{nv}{new value of parameter (starting in period of change)
(\code{nv = NA} is a special entry! see examples.)}

\item{np}{total number of periods
additional line to parameter np}
}
\description{
Creates vector with \code{np+1} entries of \code{iv}
(or optionally \code{nv} from period \code{pfc} onwards).
}
\details{
the "parameter path" refers to the series of
parameter values along all np+1 periods
}
\examples{
create_path(1, NA, NA, 5)
create_path(3, 5, 4, 10)
}
