% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{create_simulation_table}
\alias{create_simulation_table}
\title{Creates for the simulated endo. variables}
\usage{
create_simulation_table(vts, np)
}
\arguments{
\item{vts}{Vector with the abbreviation for variables that should be simulated. (vts for 'variables to simulate') The exhaustive list with all variables that can be simulated is given in the beginning of the respective simulation function.}
}
\description{
Creates an empty tibble with \code{np}+1 rows and a column for each entry in \code{vts}. See the documentation for the argument \code{vts} below.
}
\details{
This function will be used in the simulation functions to set up the grid that will then be in the said simulation function.
}
\examples{
create_simulation_table(variable_encoder(meta_BS_variables), 10)
}
