% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{docallissuetester}
\alias{docallissuetester}
\title{testerfunction}
\usage{
docallissuetester()
}
\description{
tester function to see whether the function is accessible in this function. because if so, i don't see why it would not be accessible to the doCall statement.
}
