% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{draw_transition_diagram}
\alias{draw_transition_diagram}
\title{Draw the transition diagram for a solow variant}
\usage{
draw_transition_diagram(paragrid_special, solow_variant)
}
\arguments{
\item{paragrid_special}{A special parameter grid. See the examples below to see how this input is supposed to look like.}

\item{solow_variant}{A model code such as "BS", "GS", ...}
}
\description{
Draw the transition diagram for a solow variant
}
\examples{
draw_transition_diagram(
list(B = c(1),
     alpha = 1/3, 
     delta = 0.2, 
     n = 0.005, 
     s = c(0.25, 0.05)
),
"BS"
)
}
