% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{ESSRL_SS_YpW}
\alias{ESSRL_SS_YpW}
\title{Get steady state value of YpW in the the ESSRL variant}
\usage{
ESSRL_SS_YpW(KpW, YpW, A, X, L, alpha, beta)
}
\description{
Get steady state value of YpW in the the ESSRL variant
}
