% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getCommonParameters}
\alias{getCommonParameters}
\title{Yields the union of required parameters for two solow variants}
\usage{
getCommonParameters(ModelCode1, ModelCode2)
}
\arguments{
\item{ModelCode1}{Model abbreviation for some Solow variant.}

\item{ModelCode2}{Model abbreviation for some Solow variant.}
}
\description{
Yields the union of required parameters for two solow variants
}
\examples{
getUnifiedParameters("BS", "ESSOE")
}
