% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getRequiredInputsToTE}
\alias{getRequiredInputsToTE}
\title{Get the inputs that are required to draw a models transition diagram (with the transition equation)}
\usage{
getRequiredInputsToTE(ModelCode)
}
\arguments{
\item{ModelCode}{Model abbreviation, such as "BS", "GS", "ESHC", ESSRO". (See the vignette to this package for an exhaustive listing of possible model codes.)}
}
\description{
Get the inputs that are required to draw a models transition diagram (with the transition equation)
}
\examples{
getRequiredInputsToTE("BS")
}
