% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getRequiredParams}
\alias{getRequiredParams}
\title{Get the parameters that define a certain Solow variant}
\usage{
getRequiredParams(ModelCode)
}
\arguments{
\item{ModelCode}{Model abbreviation, such as "BS", "GS", "ESHC", ESSRO". (See the vignette to this package for an exhaustive listing of possible model codes.)}
}
\description{
Each simulation function requires certain parameters to be
set. This function yields precisely those parameters
(resp. their abbreviations).
}
\examples{
getRequiredParams("BS")
}
