% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getVariablesAvailableToBeVisualised}
\alias{getVariablesAvailableToBeVisualised}
\title{to be filled}
\usage{
getVariablesAvailableToBeVisualised(ModelCode1, ModelCode2)
}
\arguments{
\item{ModelCode1}{Model abbreviation for some Solow variant.}

\item{ModelCode2}{Model abbreviation for some Solow variant.}
}
\description{
tbd
}
\details{
tbd
}
\examples{
getVariablesAvailableToBeVisualised("BS", "ESSRO")
}
