% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESEGRomer.R
\name{SimulateExtendedSolowModelEndogenousGrowthRomer}
\alias{SimulateExtendedSolowModelEndogenousGrowthRomer}
\title{Simulates the ESEGR Solow variant}
\usage{
SimulateExtendedSolowModelEndogenousGrowthRomer(paragrid, np, startvals)
}
\arguments{
\item{paragrid}{Parameter grid from \code{create_parameter_grid()}.}

\item{np}{Number of periods.}

\item{startvals}{List with starting values for the variables A, K, L, H, #' R, V. **Remark**: Use \code{getRequiredStartingValues(ModelCode)} to see which variables need to have a starting value in the respective Solow model.}
}
\description{
Simulates all (both primary and secondary) endogenous variables to the extended Solow growth model with endogenous technological growth as in the version put forward by Romer.
}
\note{
The structural equations to this model can be found in the vignette to this package:
\code{vignette("SolowVariants")}
}
