% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{simulation_correctness_checker}
\alias{simulation_correctness_checker}
\title{Check correctness by comparing simulated (endo.) variables to their steady state values}
\usage{
simulation_correctness_checker(
  last_row_simulation,
  last_row_parameter,
  solow_variant
)
}
\arguments{
\item{last_row_simulation}{Vector of the values of both primary and secondary variables in the last period of the simulation.}

\item{last_row_parameter}{Vector of the values of the exo. paramter (from the paramter grid).}

\item{solow_variant}{String indicating the model, such as "BS" or "ESHC".}
}
\description{
Compare variables in the final period of the simulation to their respective steady state value (given the exo. paramters).
}
\details{
This function presumes that the number of periods simulated were such that the final values are near steady state. If the steady state path is disrupted by a parameter change in the second to last period, this function will yield misleading results.
}
