% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{steadystate_checker}
\alias{steadystate_checker}
\title{Check correctness by comparing simulated (endo.) variables to their steady state values}
\usage{
steadystate_checker(sim_data, parameter_grid, solow_variant)
}
\arguments{
\item{sim_data}{The tibble that is being filled in the simulation function, e.g. \code{SimulateBasicSolowModel()}.}

\item{parameter_grid}{The output from \code{create_parameter_grid(...)}.}

\item{solow_variant}{String indicating the model, such as "BS" or "ESHC".}
}
\description{
Compare variables in the final period of the simulation to their respective steady state value (given the exo. paramters).
}
\details{
This function presumes that the number of periods simulated were such that the final values are near steady state. If the steady state path is disrupted by a parameter change in the second to last period, this function will yield misleading results.
}
