% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{variable_encoder}
\alias{variable_encoder}
\title{Switch between variable abbreviations and full variables names}
\usage{
variable_encoder(variables)
}
\arguments{
\item{variables}{Vector of length greater or equal to one containing the full variable name(s).}
}
\description{
Encode variables, that is, to encode the full variable variable name as defined at the top of simulation files and encode them into their respective abbreviation.
}
\examples{
variable_encoder("Total Factor Productivity")
variable_encoder(meta_BS_variables)
}
