% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{VisualiseSimulation}
\alias{VisualiseSimulation}
\title{Visualise selected variables of a simulation}
\usage{
VisualiseSimulation(simulation_data, variables, scale_identifier)
}
\arguments{
\item{simulation_data}{The output tibble of any simulation function.}

\item{variables}{The variables to visualise in *abbreviated* form.}

\item{scale_identifier}{string to indicate freely floating scales ("free") or fixed sales ("fixed")}
}
\description{
Visualise selected variables of a simulation
}
